Create signature for DLL sideloading
There may be other indicators that appear in side loading. Still currently experimenting but this seems to work fine against some cases. For this to trigger either it needs to occur naturally in the infection chain (i.e. script downloads and sideloads something) or a zip needs to be submitted with the zip_compound calling the correct file in the zip that will sideload the DLL.